chore: rename interfaces & implementations
