+++ src/main/java/ru/mirea/megatracker/interfaces/NoteService.java
@@ -1,6 +1,6 @@
 package ru.mirea.megatracker.interfaces;
 
-public interface INoteService {
+public interface NoteService {
     public void setNoteForCoin(String email, String ticker, String newNote);
     public void setFavoriteForCoin(String email, String ticker, boolean isFavorite);
 }
