+++ src/main/java/ru/mirea/megatracker/services/NoteServiceImpl.java
@@ -4,7 +4,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import ru.mirea.megatracker.interfaces.INoteService;
 import ru.mirea.megatracker.models.Note;
 import ru.mirea.megatracker.models.User;
 import ru.mirea.megatracker.repositories.NotesRepository;
@@ -14,12 +13,12 @@ import java.util.Optional;
 
 @Service
 @Transactional(readOnly = true)
-public class NoteService implements INoteService {
+public class NoteServiceImpl implements ru.mirea.megatracker.interfaces.NoteService {
     private final NotesRepository notesRepository;
     private final UsersRepository usersRepository;
 
     @Autowired
-    public NoteService(NotesRepository notesRepository, UsersRepository usersRepository) {
+    public NoteServiceImpl(NotesRepository notesRepository, UsersRepository usersRepository) {
         this.notesRepository = notesRepository;
         this.usersRepository = usersRepository;
     }
